System.Command.QQ

Description

Quasiquoters for external commands

Quasiquoters

sh :: QuasiQuoter

Quasiquoter for the default shell

"default" here means it uses value of SHELL environment variable or /bin/sh if it is not set.

>>> [sh|echo "hi!"|] :: IO ExitCode
hi!
ExitSuccess
>>> [sh|echo "hi!"|] :: IO String
"hi!\n"

Haskell values can be embedded with Ruby-like syntax:

>>> let n = 7
>>> [sh|echo "#{n} apples!"|] :: IO String
"7 apples!\n"

Works only for expressions (obviously):

>>> return 3 :: IO [sh|blah|]

<interactive>:74:16:
    Exception when trying to run compile-time code:
      this quasiquoter does not support splicing types
      Code: quoteType sh "blah"

shell :: FilePath -> QuasiQuoter

Shell commands quasiquoter maker

"Shell" here means something that implements the following interface:

 <SHELL> -c <COMMAND>

e.g. sh, bash, zsh, ksh, tcsh, python, etc

Everything that applies to sh applies to shell

interpreter :: FilePath -> QuasiQuoter

Interpreter commands quasiquoter maker

"Interpreter" here means something that implements the following interface:

 <INTERPRETER> -e <COMMAND>

e.g. perl, ruby, ghc, etc

Everything that applies to sh applies to interpreter

Customizations

quoter :: (String -> Q Exp) -> QuasiQuoter

Construct quasiquoter from function taking the string and producing Haskell expression.

Other kinds of quasiquoters (patterns, types or declarations quasiquoters) will fail in compile time

callCommand

Arguments

:: FilePath
Command path

-> [String]
Arguments that go to command before quasiquoter contents

-> String
Quasiquoter contents

-> Q Exp

Construct Haskell expression for external command call

class Eval r where

Different interesting return types for quasiquoters

Instances here mostly resemble the types of things in System.Process

System.Command.QQ.CommandT shows how to use Eval to provide nice DSL for sequencing external commands

Methods

eval :: String -> [String] -> r

class Embed a where

Embed haskell values into external commands

I recommend using -XExtendedDefaultRules for modules where you want to embed values, it would save for annoying type annotations for numeric literals

 embed . embed = embed

Methods

embed :: a -> String