System.Command.QQ.Embed

Description

Haskell values embedding

Documentation

class Embed a where

Embed haskell values into external commands

I recommend using -XExtendedDefaultRules for modules where you want to embed values, it would save for annoying type annotations for numeric literals

 embed . embed = embed

Methods

embed :: a -> String